// src/vs/platform/agentHost/node/codex/protocol/generated/v2/McpServerStatus.ts
// AUTOGENERATED - do not edit.
// Source: `codex app-server generate-ts --experimental`
// Generated from @openai/codex 0.142.0, licensed Apache-2.0.
// Modified by VS Code to rewrite imports and apply repository formatting.
// See README.md in this directory for provenance and licensing notes.
// Regenerate: npm run codex:gen-protocol
// Source file: v2/McpServerStatus.ts

import type { McpServerInfo } from "../McpServerInfo.js";
import type { Resource } from "../Resource.js";
import type { ResourceTemplate } from "../ResourceTemplate.js";
import type { Tool } from "../Tool.js";
import type { McpAuthStatus } from "./McpAuthStatus.js";

export type McpServerStatus = { name: string, serverInfo: McpServerInfo | null, tools: { [key in string]?: Tool }, resources: Array<Resource>, resourceTemplates: Array<ResourceTemplate>, authStatus: McpAuthStatus, };