// src/vs/platform/agentHost/node/codex/protocol/generated/Tool.ts
// AUTOGENERATED - do not edit.
// Source: `codex app-server generate-ts --experimental`
// Generated from @openai/codex 0.142.0, licensed Apache-2.0.
// Modified by VS Code to rewrite imports and apply repository formatting.
// See README.md in this directory for provenance and licensing notes.
// Regenerate: npm run codex:gen-protocol
// Source file: Tool.ts

import type { JsonValue } from "./serde_json/JsonValue.js";

/**
 * Definition for a tool the client can call.
 */
export type Tool = { name: string, title?: string, description?: string, inputSchema: JsonValue, outputSchema?: JsonValue, annotations?: JsonValue, icons?: Array<JsonValue>, _meta?: JsonValue, };